fix(oncall-api): return all users from lowest escalation level
Instead of return all users that match the escalation level of the first user returned in the oncall list, return users that match the lowest escalation level.
Previously, if a higher escalation level user was the first item in the array, only users at that same level would be returned.

Fixes PagerDuty/backstage-plugin#77